Projects/ReconnaissanceAI/requirements.txt
python-nmap
pywin32
pyttsx3